+++ src/main/java/com/team03/monew/dto/interest/CursorPageResponseInterestDto.java
@@ -1,13 +1,15 @@
-package com.team03.monew.dto.junhyeok;
+package com.team03.monew.dto.interest;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-public record CursorPageResponseNotificationDto(
-    List<NotificationDto> content,
+// CursorPageResponseInterestDto.java
+public record CursorPageResponseInterestDto(
+    List<InterestDto> content,
     String nextCursor,
     LocalDateTime nextAfter,
     int size,
     long totalElements,
     boolean hasNext
 ) {}
+
