+++ src/main/java/com/team03/monew/dto/notification/CursorPageResponseNotificationDto.java
@@ -1,15 +1,13 @@
-package com.team03.monew.dto.junhyeok;
+package com.team03.monew.dto.notification;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-// CursorPageResponseInterestDto.java
-public record CursorPageResponseInterestDto(
-    List<InterestDto> content,
+public record CursorPageResponseNotificationDto(
+    List<NotificationDto> content,
     String nextCursor,
     LocalDateTime nextAfter,
     int size,
     long totalElements,
     boolean hasNext
 ) {}
-
