+++ src/main/java/com/team03/monew/repository/Custom/InterestRepositoryCustom.java
@@ -1,6 +1,6 @@
-package com.team03.monew.repository.interest;
+package com.team03.monew.repository.Custom;
 
-import com.team03.monew.dto.junhyeok.CursorPageResponseInterestDto;
+import com.team03.monew.dto.interest.CursorPageResponseInterestDto;
 import java.time.LocalDateTime;
 import java.util.UUID;
