--- src/main/java/com/team03/monew/repository/interest/InterestRepository.java
@@ -1,8 +1,0 @@
-package com.team03.monew.repository.interest;
-
-import com.team03.monew.entity.Interest;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface InterestRepository extends JpaRepository<Interest, UUID>, InterestRepositoryCustom {
-}
